Added more tests to the controller

+++ code/core/src/com/mygdx/game/controller/entities/PlayerController.java
@@ -62,7 +62,7 @@ public class PlayerController extends EntityController {
 
     @Override
     public void leftWallCollision() {
-        body.applyForceToCenter((float) (100 + 5 * Math.pow((float) body.getLinearVelocity().x, 2)), 0, true);
+        body.applyForceToCenter((float) (100 + 5 * Math.pow( body.getLinearVelocity().x, 2)), 0, true);
     }
 
     @Override
